requirements.txt
aiofiles==23.2.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asyncer==0.0.8
attrs==25.1.0
audioop-lts==0.2.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmudict==1.0.32
colorama==0.4.6
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
fast-depends==2.4.12
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
gradio==5.17.1
gradio_client==1.7.1
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.27.1
numpy==2.2.3
openai==1.63.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
phidata==2.7.10
pillow==11.1.0
pronouncing==0.2.0
protobuf==5.29.3
pyarrow==19.0.1
pyautogen==0.7.5
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pywin32==308
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
ruff==0.9.7
safehttpx==0.1.6
semantic-version==2.10.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.45.3
streamlit==1.42.2
tenacity==9.0.0
termcolor==2.5.0
tiktoken==0.9.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
websockets==14.2
zipp==3.21.0
